+++ test/testutil.h
@@ -6,26 +6,26 @@
 static int vtokeq(const char *s, jsmntok_t *t, unsigned long numtok,
                   va_list ap) {
   if (numtok > 0) {
-    uint64_t i;
-    int64_t start, end, size;
-    int64_t type;
+    uint32_t i;
+    int start, end, size;
+    int type;
     char *value;
 
     size = -1;
     value = NULL;
     for (i = 0; i < numtok; i++) {
-      type = va_arg(ap, int64_t);
+      type = va_arg(ap, int);
       if (type == JSMN_STRING) {
         value = va_arg(ap, char *);
-        size = va_arg(ap, int64_t);
+        size = va_arg(ap, int);
         start = end = -1;
       } else if (type == JSMN_PRIMITIVE) {
         value = va_arg(ap, char *);
         start = end = size = -1;
       } else {
-        start = va_arg(ap, int64_t);
-        end = va_arg(ap, int64_t);
-        size = va_arg(ap, int64_t);
+        start = va_arg(ap, int);
+        end = va_arg(ap, int);
+        size = va_arg(ap, int);
         value = NULL;
       }
       if (t[i].type != type) {
@@ -49,7 +49,7 @@ static int vtokeq(const char *s, jsmntok_t *t, unsigned long numtok,
 
       if (s != NULL && value != NULL) {
         const char *p = s + t[i].start;
-        if (strlen(value) != (uint64_t)(t[i].end - t[i].start) ||
+        if (strlen(value) != (unsigned int)(t[i].end - t[i].start) ||
             strncmp(p, value, t[i].end - t[i].start) != 0) {
           printf("token %lu value is %.*s, not %s\n", i, t[i].end - t[i].start,
                  s + t[i].start, value);
